제품 검색 UI 구현
