+++ PyeonHae/PyeonHae/Utils/RoundedCorner.swift
@@ -1,0 +1,18 @@
+//
+//  RoundedCorner.swift
+//  PyeonHae
+//
+//  Created by 정건호 on 2023/02/26.
+//
+
+import SwiftUI
+
+struct RoundedCorner: Shape {
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
+        return Path(path.cgPath)
+    }
+}
