+++ PyeonHae/PyeonHae/Views/Review/EditReview/EditReviewView.swift
@@ -31,6 +31,7 @@ struct EditReviewView: View {
             .background(Color.white)
             .toast(message: "사진은 최대 3장까지 추가할 수 있습니다.", isShowing: $showToast, config: .init())
         }
+        .blur(radius: showSearchProductView ? 2 : 0)
         .bottomSheet(isPresented: $showSearchProductView) {
             SearchProductView(showSearchProductView: $showSearchProductView)
         }
